formatted lesson 4
